Added --always-copy to virtualenv call
This make it so that the files in the vent are copied from the source 
machine when creating the jar files. If this is not present, recent versions 
of virtualenv will symlink and cause bad links in the jar files. Bad links 
mean that the files will not exist when the jar is actually run. This fix 
ensures that all required files in the venv are copied over so you do not 
actually have to have python installed on your pyleus cluster.

This is the root cause of the `ImportError: No module named 'encodings'` problem in #50